% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-get.R
\name{as_auth}
\alias{as_auth}
\title{Coerce to auth function}
\usage{
as_auth(auth)
}
\arguments{
\item{auth}{Object coerced to auth function.
\itemize{
\item If \code{NULL}, returns default auth-function returned by \code{bx_auth()}.
\item If \code{bx_auth()} returns \code{NULL}, tries \code{\link[=bx_auth_get]{bx_auth_get()}}.
}}
}
\description{
Coerce to auth function
}
\keyword{internal}
