% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-serialize-parse.R
\name{auth_serialize}
\alias{auth_serialize}
\title{Serialize an auth function}
\usage{
auth_serialize(auth)
}
\arguments{
\item{auth}{can be:
\itemize{
\item \code{function} with S3 class \code{cardboard_auth}: uses \code{auth}.
\item \code{character}: uses auth function using name of stored in keyring,
see \code{bx_keyring_list()}.
\item \code{list}: uses \code{\link[=bx_auth_parse]{bx_auth_parse()}}.
\item \code{NULL}: if \code{bx_keyring_list()} has exactly one entry, uses it.
}}
}
\value{
Named \code{list} with elements \code{type}, \code{params}.
Returned invisibly because it contains secret information.
}
\description{
Serialize an auth function
}
\keyword{internal}
