% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-coerce.R
\name{bx_auth}
\alias{bx_auth}
\title{Coerce to an auth function}
\usage{
bx_auth(auth)
}
\arguments{
\item{auth}{can be:
\itemize{
\item \code{function} with S3 class \code{cardboard_auth}: uses \code{auth}.
\item \code{character}: uses auth function using name of stored in keyring,
see \code{bx_keyring_list()}.
\item \code{NULL}: uses default auth-function.
}}
}
\value{
\code{function} with S3 class \code{cardboard_auth}, fully operational
drop-in replacement for \code{httr2::req_oauth_auth_code()} or
\code{httr2::req_oauth_auth_credentials()}.
}
\description{
Coerce to an auth function
}
