% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-to-from-list.R
\name{bx_auth_from_list}
\alias{bx_auth_from_list}
\title{Create an auth functiuon}
\usage{
bx_auth_from_list(list_auth, .test = FALSE)
}
\arguments{
\item{list_auth}{\code{list}}

\item{.test}{\code{logical} indicates to test by calling the \code{users/me} resource.}
}
\value{
\code{function} that takes only an \code{httr2::request} object and returns a
\code{httr2::request} object. This function can be used place of
\code{httr2::req_oauth_auth_code()} or \code{httr2::req_oauth_auth_credentials()}.
}
\description{
Create an auth functiuon
}
\keyword{internal}
