% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-get.R
\name{bx_auth_get}
\alias{bx_auth_get}
\title{Get auth function}
\usage{
bx_auth_get(auth)
}
\arguments{
\item{auth}{can be:
\itemize{
\item \code{function} with S3 class \code{cardboard_auth}: uses \code{auth}.
\item \code{character}: uses auth function using name of stored in keyring,
see \code{bx_keyring_list()}.
\item \code{list}: uses \code{\link[=bx_auth_from_list]{bx_auth_from_list()}}.
\item \code{NULL}: if \code{bx_keyring_list()} has exactly one entry, uses it.
}}
}
\value{
\code{function} with S3 class \code{cardboard_auth}, fully operational
drop-in replacement for \code{httr2::req_oauth_auth_code()} or
\code{httr2::req_oauth_auth_credentials()}.
}
\description{
Get auth function
}
