% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-test-inspect.R
\name{bx_auth_inspect}
\alias{bx_auth_inspect}
\title{Inspect an auth function}
\usage{
bx_auth_inspect(auth = NULL)
}
\arguments{
\item{auth}{Object coerced to auth function.
\itemize{
\item If \code{NULL}, returns default auth-function returned by \code{bx_auth()}.
\item If \code{bx_auth()} returns \code{NULL}, tries \code{\link[=bx_auth_get]{bx_auth_get()}}.
}}
}
\value{
\code{function} with S3 class \code{cardboard_auth}, fully operational
drop-in replacement for \code{httr2::req_oauth_auth_code()} or
\code{httr2::req_oauth_auth_credentials()}.
}
\description{
Inspect an auth function
}
