% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-make.R
\name{bx_auth_make_code}
\alias{bx_auth_make_code}
\alias{bx_auth_make_credentials}
\title{Make an auth function}
\usage{
bx_auth_make_code(id = NULL, secret = NULL, cache_disk = FALSE, .test = TRUE)

bx_auth_make_credentials(
  id = NULL,
  secret = NULL,
  subject_type = c("enterprise", "user"),
  subject_id = NULL,
  cache_disk = FALSE,
  .test = TRUE
)
}
\arguments{
\item{id}{\code{character} client id, taken from Box app.}

\item{secret}{\code{character} client secret, taken from Box app.}

\item{cache_disk}{\code{logical} for code-based auth, indicates if the token
should be cached on disk for use in other R sessions.}

\item{.test}{\code{logical} indicates to test by calling the \code{users/me} resource.}

\item{subject_type}{\code{character} for credentials-based auth, the type of
account to use.}

\item{subject_id}{\code{character} for credentials-based auth, the id of the
account to use.}
}
\value{
\code{function} that takes only an \code{httr2:: request} object and returns a
\code{httr2:: request} object. This function can be used place of
\code{httr2::req_oauth_auth_code()} or \code{httr2::req_oauth_auth_credentials()}.
}
\description{
\itemize{
\item RMarkdown files are generally rendered in their own R session. By caching
the token, you could first authenticate to Box interactively.
\item if you use the default (\code{NULL}) for \code{id} and \code{secret}, you will be asked
interactively. This may be preferable because the \code{id} and \code{secret} will not
be saved to your command history.
\item \code{"user"} does not seem to work for credentials-based auth.
}
}
