% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-parse.R
\name{bx_auth_parse}
\alias{bx_auth_parse}
\title{Create an auth functiuon}
\usage{
bx_auth_parse(type = c("interactive", "credentials"), params, .test = FALSE)
}
\arguments{
\item{type}{\code{character}}

\item{params}{\code{list}}

\item{.test}{\code{logical} indicates to test by calling the \code{users/me} resource.}
}
\value{
\code{function} that takes only an \code{httr2::request} object and returns a
\code{httr2::request} object. This function can be used place of
\code{httr2::req_oauth_auth_code()} or \code{httr2::req_oauth_auth_credentials()}.
}
\description{
Create an auth functiuon
}
