% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-serialize.R
\name{bx_auth_serialize}
\alias{bx_auth_serialize}
\title{Serialize an auth function}
\usage{
bx_auth_serialize(auth)
}
\arguments{
\item{auth}{can be:
\itemize{
\item \code{function} with S3 class \code{cardboard_auth}: uses \code{auth}.
\item \code{character}: uses auth function using name of stored in keyring,
see \code{bx_keyring_list()}.
\item \code{NULL}: uses default auth-function.
}}
}
\value{
Named \code{list} with elements \code{type}, \code{params}.
}
\description{
Serialize an auth function
}
