% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-create.R
\name{bx_auth_test}
\alias{bx_auth_test}
\title{Test an auth function}
\usage{
bx_auth_test(function_req_auth)
}
\arguments{
\item{function_req_auth}{\code{function} created using \code{bx_auth_create_code()} or
\code{bx_auth_create_credentials()}}
}
\value{
Invisible \code{logical}, indicating success
}
\description{
Test an auth function
}
