% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-to-from-list.R
\name{bx_auth_to_list}
\alias{bx_auth_to_list}
\title{to_list an auth function}
\usage{
bx_auth_to_list(auth)
}
\arguments{
\item{auth}{Object coerced to auth function.
\itemize{
\item If \code{NULL}, returns default auth-function returned by \code{bx_auth()}.
\item If \code{bx_auth()} returns \code{NULL}, tries \code{\link[=bx_auth_get]{bx_auth_get()}}.
}}
}
\value{
Named \code{list} with elements \code{type}, \code{params}.
Returned invisibly because it contains secret information.
}
\description{
to_list an auth function
}
\keyword{internal}
