% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-encrypt-decrypt.R
\name{bx_auth_encrypt_file}
\alias{bx_auth_encrypt_file}
\alias{bx_auth_decrypt_file}
\title{Auth file encryprition}
\usage{
bx_auth_encrypt_file(auth, file, key = NULL, nonce = NULL)

bx_auth_decrypt_file(file, key, nonce = NULL, .test = TRUE)
}
\arguments{
\item{auth}{Object coerced to auth function.
\itemize{
\item If \code{NULL}, returns default auth-function returned by \code{bx_auth()}.
\item If \code{bx_auth()} returns \code{NULL}, tries \code{\link[=bx_auth_get]{bx_auth_get()}}.
}}

\item{file}{\code{character}}

\item{key}{\code{character}}

\item{nonce}{\code{character}}
}
\value{
Invisible \code{auth}, called for side effects.
}
\description{
Auth file encryprition
}
