% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-serialize-parse.R
\name{bx_auth_serialize}
\alias{bx_auth_serialize}
\title{Serialize an auth function}
\usage{
bx_auth_serialize(auth)
}
\arguments{
\item{auth}{Object coerced to auth function.
\itemize{
\item If \code{NULL}, returns default auth-function returned by \code{bx_auth()}.
\item If \code{bx_auth()} returns \code{NULL}, tries \code{\link[=bx_auth_get]{bx_auth_get()}}.
}}
}
\value{
Named \code{list} with elements \code{type}, \code{params}.
Returned invisibly because it contains secret information.
}
\description{
Serialize an auth function
}
