% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-test-inspect.R
\name{bx_auth_test}
\alias{bx_auth_test}
\title{Test an auth function}
\usage{
bx_auth_test(auth = NULL)
}
\arguments{
\item{auth}{Object coerced to auth function.
\itemize{
\item If \code{NULL}, returns default auth-function returned by \code{bx_auth()}.
\item If \code{bx_auth()} returns \code{NULL}, tries \code{\link[=bx_auth_get]{bx_auth_get()}}.
}}
}
\value{
Invisible \code{logical}, indicating success
}
\description{
Test an auth function
}
