% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-make.R
\name{bx_auth_test}
\alias{bx_auth_test}
\title{Test an auth function}
\usage{
bx_auth_test(function_req_auth)
}
\arguments{
\item{function_req_auth}{\code{function} created using \code{bx_auth_make_code()} or
\code{bx_auth_make_credentials()}}
}
\value{
\code{function_req_auth}, invisibly; called for side-effects.
}
\description{
Test an auth function
}
